Added edit functionality
DatabaseManager: new getRule(String ruleName) method to get a rule for
a given rule name. new editRule(String oldrulename, Rule newRule)
updates the rule with the given name in the DB to be the new rule.
editRule returns the widgetID so addRule can broadcast a widget update
if needed.
AddRule: Changed add button to save button. If the activity is launched
by an edit intent, a progress bar will pop while the fields are loaded
with the data of the rule. Also added new functionality to the save
button to match the new editing function, which also calls for widget
update if the rule has a widget && it’s name is changed.
AddRuleActivity.xml: Changed the parent layout to RelativeLayout, added
a children relative layout to hold the progress bar, and the other
children is a scrollview holding the linear layout filled with the
fields
progress.xml: This is completely new, the layout of the progress bar
currently only used by the addrule activity